完成影院服务开发，并测试接口 添加配置中心（未启用） 添加服务监控（Monitor未启用）

+++ OSCP-FilmService/src/main/java/org/joker/oscp/film/entity/CatDict.java
@@ -1,8 +1,8 @@
 package org.joker.oscp.film.entity;
 
-import com.baomidou.mybatisplus.annotation.TableField;
-import com.baomidou.mybatisplus.annotation.TableId;
-import com.baomidou.mybatisplus.annotation.TableName;
+import com.baomidou.mybatisplus.annotations.TableField;
+import com.baomidou.mybatisplus.annotations.TableId;
+import com.baomidou.mybatisplus.annotations.TableName;
 import lombok.Data;
 
 /**
@@ -16,7 +16,7 @@ public class CatDict {
      * 主键编号
      */
     @TableId
-    private Integer uuid;
+    private Long uuid;
     /**
      * 显示名称
      */
